Move `beforeEach` into reducer_support.ts. NEXT: Testing
