+++ frontend/resources/reducer.ts
@@ -7,12 +7,13 @@ import {
   indexRemove,
   initResourceReducer,
   afterEach,
+  beforeEach
 } from "./reducer_support";
 import { TaggedResource, SpecialStatus } from "farmbot";
 import { Actions } from "../constants";
 import { EditResourceParams } from "../api/interfaces";
-import { defensiveClone, equals, unpackUUID } from "../util";
-import { merge, get } from "lodash";
+import { defensiveClone, equals } from "../util";
+import { merge } from "lodash";
 import { SyncBodyContents } from "../sync/actions";
 import { GeneralizedError } from "./actions";
 import { initialState as helpState } from "../help/reducer";
@@ -21,8 +22,6 @@ import { farmwareState } from "../farmware/reducer";
 import { initialState as regimenState } from "../regimens/reducer";
 import { initialState as sequenceState } from "../sequences/reducer";
 import { initialState as alertState } from "../messages/reducer";
-import { warning } from "../toast/toast";
-import { ReduxAction } from "../redux/interfaces";
 
 export const emptyState = (): RestResources => {
   return {
@@ -80,47 +79,7 @@ export const emptyState = (): RestResources => {
 /** Responsible for all RESTful resources. */
 export let resourceReducer =
   generateReducer<RestResources>(emptyState())
-    .beforeEach((state, action, handler) => {
-      const { byKind, references } = state.index;
-      const w = references[Object.keys(byKind.WebAppConfig)[0]];
-      const readOnly = w &&
-        w.kind == "WebAppConfig" &&
-        w.body.user_interface_read_only_mode;
-      if (!readOnly) {
-        return handler(state, action);
-      }
-      const fail = (place: string) => {
-        console.log(action.type);
-        console.dir(action);
-        warning(`(${place}) Can't modify account data when in read-only mode.`);
-      };
-      const { kind } = unpackUUID(get(action, "payload.uuid", "x.y.z"));
-
-      switch (action.type) {
-        case Actions.EDIT_RESOURCE:
-          if (kind === "WebAppConfig") {
-            // User is trying to exit read-only mode.
-            return handler(state, action);
-          } else {
-            fail("1");
-            return state;
-          }
-        case Actions.SAVE_RESOURCE_START:
-          if (kind !== "WebAppConfig") {
-            // User is trying to make HTTP requests.
-            fail("3");
-          }
-          // User is trying to exit read-only mode.
-          return handler(state, action);
-        case Actions.BATCH_INIT:
-        case Actions.INIT_RESOURCE:
-        case Actions.OVERWRITE_RESOURCE:
-          fail("2");
-          return state;
-        default:
-          return handler(state, action);
-      }
-    })
+    .beforeEach(beforeEach)
     .afterEach(afterEach)
     .add<TaggedResource>(Actions.SAVE_RESOURCE_OK, (s, { payload }) => {
       indexUpsert(s.index, [payload], "ongoing");
